src/test/java/com/leetcode/dp/medium/CombinationSumIV_377Test.java
package com.leetcode.dp.medium;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class CombinationSumIV_377Test {

  private final CombinationSumIV_377 solution = new CombinationSumIV_377();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(arguments(new int[] {1, 2, 3}, 4, 7));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void coinChange(int[] coins, int target, int expectedResult) {
    int actualResult = solution.combinationSum4(coins, target);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
